Add files via upload
These files pertain to the sign language representation of two, three, four, and six-digit numbers in Iranian sign language. The labels for the numbers are provided in a separate file. For two-digit numbers, the end index of the first digit and the start index of the second digit are included in one file. For three-digit numbers, the start and end indices for the three digits in each sample are provided in a single file. The indices corresponding to the start and end of the digits are specified for the first 430 cases of the four-digit dataset. The indices for samples after the first 430 in the four-digit dataset, as well as the indices for the six-digit dataset, are not included here, as this data is solely for testing the dataset, and the ultimate goal is number recognition, not the indices. It is important to note that six-digit numbers actually consist of seven digits since after expressing the first three digits, the thousands are indicated, followed by the last three digits.

The classes present in this dataset include independent numbers from 1 to 9 (static numbers), 10 to 19 (beginning of numbers with dynamic movement), 20 to 90, 100 to 900, and 1000 to 9000. These are independent and single gestures; for instance, from 21 to 29 requires two gestures, whereas the number 30 is a single independent movement.